+++ src/scame/gnus/gnus-gmail.el
@@ -1,6 +1,6 @@
 ;;; gnus-gmail.el --- Gmail configuration for Gnus
 
-;; Copyright (C) 2014 Nicolas Lamirault <nicolas.lamirault@gmail.com>
+;; Copyright (C) 2014, 2015 Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 ;; This program is free software: you can redistribute it and/or modify
 ;; it under the terms of the GNU General Public License as published by
@@ -26,22 +26,35 @@
 (require 'scame-bbdb)
 
 
-;; Gmail configuration
-;; --------------------
+;; GMail IMAP configuration
+;; -------------------------
 
 ;; Place a line like the following in ~/.authinfo
 ;; machine imap.gmail.com login your-name@gmail.com password your-password port 993
 ;; and make sure that no-one else can read it with
 ;; chmod 600 ~/.authinfo
 
-(setq gnus-select-method
-      '(nnimap "gmail"
-               (nnimap-address "imap.gmail.com")
-               (nnimap-server-port 993)
-               (nnimap-stream ssl)
-               ;; press 'E' to expire email
-               (nnmail-expiry-target "nnimap+gmail:[Gmail]/Trash")
-               ))
+(defun scame--gmail-imap ()
+  "Setup gnus-select-method with IMAP GMail server."
+  (setq gnus-select-method
+        '(nnimap "gmail"
+                 (nnimap-address "imap.gmail.com")
+                 (nnimap-server-port 993)
+                 (nnimap-stream ssl)
+                 ;; press 'E' to expire email
+                 (nnmail-expiry-target "nnimap+gmail:[Gmail]/Trash")))
+  (define-key gnus-summary-mode-map
+    (kbd "B d")
+    (lambda ()
+      (interactive)
+      (gnus-summary-move-article nil "nnimap+gmail:[Gmail]/Trash")))
+  (define-key gnus-summary-mode-map
+    (kbd "B s")
+    (lambda ()
+      (interactive)
+      (gnus-summary-move-article nil "nnimap+gmail:[Gmail]/Spam"))))
+
+;; GMail SMTP configuration
 
 ;; If you don't want to be prompted for a password on every mail sent,
 ;; you can add the following line to your ~/.authinfo.
@@ -57,17 +70,6 @@
 
 ;; Keybindings
 
-(define-key gnus-summary-mode-map
-  (kbd "B d")
-  (lambda ()
-    (interactive)
-    (gnus-summary-move-article nil "nnimap+gmail:[Gmail]/Trash")))
-
-(define-key gnus-summary-mode-map
-  (kbd "B s")
-  (lambda ()
-    (interactive)
-    (gnus-summary-move-article nil "nnimap+gmail:[Gmail]/Spam")))
 
 
 (provide 'gnus-gmail)