+++ src/scame/packages/91_orga.el
@@ -37,7 +37,9 @@
 ;; Calendar
 ;; ----------
 
-(defvar french-holiday
+;; From Emacs Jd setup
+
+(defvar calendar-legal-holidays
   '((holiday-fixed 1 1 "Jour de l'an")
     (holiday-fixed 5 1 "Fête du travail")
     (holiday-fixed 5 8 "Victoire 45")
@@ -48,17 +50,28 @@
     (holiday-fixed 12 25 "Noël")
     (holiday-easter-etc 1 "Lundi de Pâques")
     (holiday-easter-etc 39 "Ascension")
-    (holiday-easter-etc 50 "Lundi de Pentecôte")
-    ;; Misc
-    (holiday-fixed 3 17 "St. Patrick's Day")
-    (holiday-float 5 0 -1 "Fête des Mères")
-    (holiday-float 6 0 3 "Fête des Pères")
-    ))
+    (holiday-easter-etc 50 "Lundi de Pentecôte")))
 
+(setq calendar-celebration-holidays
+      '((holiday-fixed 2 2 "Chandeleur")
+        (holiday-fixed 2 14 "Saint Valentin")
+        (holiday-float 3 0 1 "Fête des grands-mères")
+        (holiday-fixed 3 17 "St. Patrick's Day")
+        (holiday-fixed 4 1 "April Fools' Day")
+        (holiday-float 5 0 -1 "Fête des Mères")
+        (holiday-float 6 0 3 "Fête des Pères")
+        (holiday-fixed 6 21 "Fête de la musique")
+        (holiday-fixed 10 31 "Halloween")
+        (holiday-easter-etc -47 "Mardi Gras")))
+
+(setq calendar-holidays
+      `(,@holiday-solar-holidays
+        ,@calendar-legal-holidays
+        ,@calendar-celebration-holidays))
 
 (use-package calendar
   :config (setq calendar-date-style 'european
-                calendar-holidays french-holiday
+                calendar-holidays calendar-legal-holidays
                 calendar-mark-holidays-flag t
                 calendar-week-start-day 1 ;; week starts on monday
                 calendar-set-date-style 'iso
@@ -72,11 +85,35 @@
                 calendar-intermonth-header (propertize "Wk"
                                                        'font-lock-face 'font-lock-keyword-face)))
 
+;; Appointment
+;; ------------
+
+(defun scame--appt-disp-window (min-to-app new-time msg)
+  "Wrapper for appt popup display."
+  (notifications-notify :title (format "Appt in %s minute(s)" min-to-app)
+                        :body msg
+                        :app-name "Emacs: Org"))
+
+
+(use-package appt
+  :config (progn
+            (setq appt-audible nil
+                  appt-display-diary nil
+                  appt-display-mode-line t     ;; show in the modeline
+                  appt-message-warning-time 60 ;; warn 60 min in advance
+                  appt-display-interval 5
+                  appt-display-duration 30
+                  appt-display-format 'window
+                  appt-disp-window-function (function scame--appt-disp-window))
+            (when (require 'sauron nil t)
+              (add-to-list 'sauron-modules 'sauron-org)))
+  :init (appt-activate))
+
+
 (use-package org
   :config (progn
-	    (setq org-directory (concat user-home-directory "Org"))
+	    (setq org-directory (f-join user-home-directory "Org"))
 	    (setq org-agenda-files (list org-directory))
-
 
 	    ;; Links
 	    ;; ------
@@ -88,7 +125,7 @@
 	    ;; Tasks
 	    ;; ------
 
-	    (setq org-default-notes-file "~/Org/diary.org")
+	    (setq org-default-notes-file (concat org-directory "/diary.org"))
 	    (setq org-todo-keywords
 		  (quote ((sequence "TODO(t)" "NEXT(n)" "|" "DONE(d)")
 			  (sequence "WAITING(w@/!)" "HOLD(h@/!)" "|"
@@ -125,7 +162,10 @@
 			   "* PHONE %? :PHONE:\n%U" :clock-in t :clock-resume t)
 			  )))
 
-	    (setq org-clock-persist 'history)
+	    (setq org-log-done t
+                  org-clock-persist t
+                  org-clock-out-when-done nil)
+
 	    (org-clock-persistence-insinuate)
 
 	    (setq org-use-fast-todo-selection t)
@@ -133,13 +173,21 @@
 
 	    ;; Agenda views
 	    ;; -------------
-
 	    ;; Do not dim blocked tasks
 	    (setq org-agenda-dim-blocked-tasks nil)
-
 	    ;; Compact the block agenda view
 	    (setq org-agenda-compact-blocks t)
+            (setq org-agenda-clockreport-parameter-plist
+                  '(:link nil :maxlevel 4 :emphasize t))
 
+            ;; Appointment
+            ;; ------------
+            (org-agenda-to-appt)
+            (add-hook 'org-finalize-agenda-hook 'org-agenda-to-appt)
+            (add-hook 'org-remember-after-finalize-hook 'org-agenda-to-appt)
+            (add-hook 'org-mode-hook (lambda()
+                                       (add-hook 'before-save-hook
+                                                 'org-agenda-to-appt t)))
 
 	    ;; Clock Setup
 	    ;; ------------
@@ -171,8 +219,9 @@
 	 ("C-c o a" . org-agenda)
 	 ("C-c o b" . org-iswitchb)))
 
-	    ;; Exporting
-	    ;; -----------
+
+;; Exporting
+;; -----------
 
 
 ;; Explicitly load required exporters